Fix seven bugs found by the Phase 2 documentation review
Writing the contract documents meant reading model/ closely, which turned
up seven defects. Each was reproduced before being fixed, and each fix
carries a regression test verified to FAIL against the code as it stood.

F-6 -- run_model.py --rel-tol and --abs-tol crashed on every use. argparse
declared them type=float; Tolerance then calls D(), which refuses floats
by design (1.15, 4.4). The defaults are Decimal constants so they passed
through, and the failure appeared only when a user supplied the flag the
README documents as the supported override. It printed the entire report
and then died on an uncaught exception, because PrecisionError derives
from ValueError and main() caught only ProvenanceError. Both flags now
parse as strings, which is what 4.2 asks for anyway, and an unparseable
value exits 2 instead of a traceback.

F-7a -- sensitivity._shift_wacc silently missed its target when beta was
zero. With beta zero, CAPM gives cost of equity = risk-free rate whatever
the premium, so no ERP reaches the target; the code fell back to ZERO and
returned an unshifted cost of capital, while the grid cell was still
labelled with the target WACC it did not have. It now refuses and says
why.

F-7b -- tv_share_of_ev returns None rather than dividing by zero (17.27),
and its only consumer formatted it with :.1%, which raises TypeError on
None. The guard converted a division error into a formatting error.

F-7c -- dividends used silent precedence where every comparable driver
refuses. _pick_driver rejects declaring both a percentage and an absolute
for COGS, opex and CapEx; dividends bypassed it and quietly preferred
dividends_amount over dividend_payout_ratio. Now rejected like the rest.

F-10a -- two of the twelve STEP 37 checks called the same function with
identical arguments, so the panel reported twelve results from eleven
tests. They are now distinct: "Forecast cash flow reconciliation" checks
that each cash-flow subtotal equals the items beneath it, and "Ending
cash linkage" checks that the cash balance follows from those subtotals.
A corrupted balance sheet now fails the second and passes the first,
which is the point of having both.

F-10b -- "FCFF matches three-statement forecast" compared build_fcff's
output against fcff_inputs, the same function build_fcff used to produce
it, so it verified only that nothing mutated in between. It now rebuilds
each term from the statements and schedules directly and recomputes FCFF
from them.

Ledger.cross_check was implemented, tested, and called by nothing, so a
filing whose own subtotal disagreed with its line items never reached the
panel. Wired in as a thirteenth check, labelled 17.10 because STEP 37
does not list it. Specification 4.10's relative_error now appears in the
discrepancy message, where it was previously also unused.

Also: numeric.quantize_for_display is now the display boundary report.py
rounds through (4.9, 4.18) rather than bare f-strings; two annotations
that drifted in the Decimal port are corrected; and Units.multiplier is
documented as correct-but-unused, with its values pinned by a test,
because 4.6 unit normalization is genuinely unimplemented and will matter
the moment 2.3.a permits a second document.

69 tests, green on Python 3.10 through 3.13. The panel now reports 13.

+++ model/loader.py
@@ -176,7 +176,7 @@ def load_historical(path: Path, periods: Periods) -> dict[Statement, Ledger]:
 
 
 # --- STEP 10-11 / 16 ------------------------------------------------------
-def load_assumptions(path: Path, periods: Periods) -> tuple[Assumptions, TaxSchedule, dict[str, float]]:
+def load_assumptions(path: Path, periods: Periods) -> tuple[Assumptions, TaxSchedule, dict[str, Decimal]]:
     data = _load_yaml(path)
     assumptions = Assumptions()
